+++ concordium-node/src/rpc.rs
@@ -10,7 +10,6 @@ use crate::{
         helpers::{ConsensusFfiResponse, ConsensusIsInBakingCommitteeResponse, PacketType},
         messaging::{ConsensusMessage, MessageType},
     },
-    failure::Fallible,
     network::NetworkId,
     p2p::{bans::PersistedBanId, P2PNode},
     read_or_die,
@@ -43,7 +42,7 @@ impl RpcServerImpl {
         node: Arc<P2PNode>,
         consensus: Option<ConsensusContainer>,
         conf: &configuration::RpcCliConfig,
-    ) -> Fallible<Self> {
+    ) -> anyhow::Result<Self> {
         let listen_addr =
             SocketAddr::from((IpAddr::from_str(&conf.rpc_server_addr)?, conf.rpc_server_port));
 
@@ -56,7 +55,7 @@ impl RpcServerImpl {
     }
 
     /// Starts the gRPC server.
-    pub async fn start_server(&mut self) -> Fallible<()> {
+    pub async fn start_server(&mut self) -> anyhow::Result<()> {
         let self_clone = self.clone();
         let server = Server::builder().add_service(P2pServer::new(self_clone));
 
@@ -379,22 +378,14 @@ impl P2p for RpcServerImpl {
         }))
     }
 
+    #[allow(deprecated)] // this is allowed until we remove the staging_net_username from the RPC
+                         // specification.
     async fn node_info(&self, req: Request<Empty>) -> Result<Response<NodeInfoResponse>, Status> {
         authenticate!(req, self.access_token);
         let node_id = Some(self.node.id().to_string());
         let peer_type = self.node.peer_type().to_string();
         let current_localtime =
             SystemTime::now().duration_since(UNIX_EPOCH).expect("Time went backwards").as_secs();
-        let staging_net_username = {
-            #[cfg(feature = "staging_net")]
-            {
-                Some(self.node.config.staging_net_username.clone())
-            }
-            #[cfg(not(feature = "staging_net"))]
-            {
-                None
-            }
-        };
         Ok(Response::new(match self.consensus {
             Some(ref consensus) => {
                 let consensus_baking_committee_status = consensus.in_baking_committee();
@@ -421,13 +412,13 @@ impl P2p for RpcServerImpl {
                         }
                     },
                     consensus_finalizer_committee: consensus.in_finalization_committee(),
-                    staging_net_username,
                     consensus_baker_id: match consensus_baking_committee_status {
                         ConsensusIsInBakingCommitteeResponse::ActiveInCommittee(baker_id) => {
                             Some(baker_id)
                         }
                         _ => None,
                     },
+                    staging_net_username: None,
                 }
             }
             None => NodeInfoResponse {
@@ -440,8 +431,8 @@ impl P2p for RpcServerImpl {
                 consensus_baker_committee: node_info_response::IsInBakingCommittee::NotInCommittee
                     as i32,
                 consensus_finalizer_committee: false,
-                staging_net_username,
                 consensus_baker_id: None,
+                staging_net_username: None,
             },
         }))
     }
@@ -798,8 +789,9 @@ impl P2p for RpcServerImpl {
     #[cfg(not(feature = "network_dump"))]
     async fn dump_start(
         &self,
-        _req: Request<DumpRequest>,
+        req: Request<DumpRequest>,
     ) -> Result<Response<BoolResponse>, Status> {
+        authenticate!(req, self.access_token);
         warn!("DumpStart RPC request received, but the \"network_dump\" feature is not active");
         Err(Status::new(Code::Unavailable, "Feature \"network_dump\" is not active"))
     }
@@ -828,7 +820,8 @@ impl P2p for RpcServerImpl {
     }
 
     #[cfg(not(feature = "network_dump"))]
-    async fn dump_stop(&self, _req: Request<Empty>) -> Result<Response<BoolResponse>, Status> {
+    async fn dump_stop(&self, req: Request<Empty>) -> Result<Response<BoolResponse>, Status> {
+        authenticate!(req, self.access_token);
         warn!("DumpStop RPC request received, but the \"network_dump\" feature is not active");
         Err(Status::new(Code::Unavailable, "Feature \"network_dump\" is not active"))
     }
@@ -854,7 +847,6 @@ mod tests {
         },
     };
     use chrono::prelude::Utc;
-    use failure::Fallible;
     use tonic::{metadata::MetadataValue, transport::channel::Channel, Code, Request};
 
     use grpc_api::p2p_client::P2pClient;
@@ -866,7 +858,7 @@ mod tests {
     // The intended use is for spawning nodes for testing gRPC api.
     async fn create_test_rpc_node(
         nt: PeerType,
-    ) -> Fallible<(P2pClient<Channel>, Arc<P2PNode>, DeletePermission)> {
+    ) -> anyhow::Result<(P2pClient<Channel>, Arc<P2PNode>, DeletePermission)> {
         let (node, dp) =
             make_node_and_sync(next_available_port(), vec![100], nt, dummy_regenesis_blocks())
                 .unwrap();
@@ -889,7 +881,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_grpc_noauth() -> Fallible<()> {
+    async fn test_grpc_noauth() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
 
         match client.peer_version(req_with_auth!(grpc_api::Empty {}, "derp")).await {
@@ -901,7 +893,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_peer_version() -> Fallible<()> {
+    async fn test_peer_version() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         assert_eq!(
             client
@@ -917,7 +909,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_peer_uptime() -> Fallible<()> {
+    async fn test_peer_uptime() -> anyhow::Result<()> {
         let t0 = Utc::now().timestamp_millis() as u64;
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
 
@@ -936,7 +928,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_peer_total_received() -> Fallible<()> {
+    async fn test_peer_total_received() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         let port = next_available_port();
         let (node2, dp2) =
@@ -956,7 +948,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_peer_total_sent() -> Fallible<()> {
+    async fn test_peer_total_sent() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         let port = next_available_port();
         let (node2, dp2) =
@@ -976,7 +968,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_peer_connect() -> Fallible<()> {
+    async fn test_peer_connect() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         let port = next_available_port();
         let (node2, dp2) =
@@ -1003,7 +995,7 @@ mod tests {
     // `tests/consensus-tests.rs`
 
     #[tokio::test]
-    async fn test_join_network() -> Fallible<()> {
+    async fn test_join_network() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         let port = next_available_port();
         let (node2, dp2) =
@@ -1024,7 +1016,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_leave_network() -> Fallible<()> {
+    async fn test_leave_network() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         let port = next_available_port();
         let (node2, dp2) =
@@ -1045,7 +1037,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_peer_stats() -> Fallible<()> {
+    async fn test_peer_stats() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         let port = next_available_port();
         let (node2, dp2) =
@@ -1069,7 +1061,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_peer_list() -> Fallible<()> {
+    async fn test_peer_list() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         let port = next_available_port();
         let (node2, dp2) =
@@ -1099,7 +1091,7 @@ mod tests {
 
     #[cfg(flag = "bootstrapper")]
     #[tokio::test]
-    async fn test_grpc_peer_list_bootstrapper() -> Fallible<()> {
+    async fn test_grpc_peer_list_bootstrapper() -> anyhow::Result<()> {
         let (mut client, _, dp) = create_test_rpc_node(PeerType::Bootstrapper).await.unwrap();
         let req = req_with_auth!(
             grpc_api::PeersRequest {
@@ -1113,7 +1105,7 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_node_info() -> Fallible<()> {
+    async fn test_node_info() -> anyhow::Result<()> {
         let instant1 = (Utc::now().timestamp_millis() as u64) / 1000;
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         let reply = client.node_info(req_with_auth!(grpc_api::Empty {}, TOKEN)).await.unwrap();
@@ -1143,7 +1135,7 @@ mod tests {
     // - Get last final instance info
 
     #[tokio::test]
-    async fn test_shutdown() -> Fallible<()> {
+    async fn test_shutdown() -> anyhow::Result<()> {
         let (mut client, node, dp) = create_test_rpc_node(PeerType::Node).await.unwrap();
         assert!(
             client